--- rakiyama/check_envp.c
@@ -1,11 +1,0 @@
-# include <stdio.h>
-
-int main(int ac, char *av[], char *envp[])
-{
-	while (*envp)
-	{
-		printf("%s\n", *envp);
-		envp++;
-	}
-	return (0);
-}
